python-api/requirements.txt
flask==3.0.2
redis==5.0.1
requests==2.31.0
mysql-connector-python==8.1.0